amap_create: remove unnecessary cast
YT: RTOS-1237

+++ vm/amap.c
@@ -137,7 +137,7 @@ amap_t *amap_create(amap_t *amap, size_t *offset, size_t size)
 	}
 
 	(void)proc_lockInit(&new->lock, &proc_lockAttrDefault, "amap.map");
-	new->size = (unsigned int)i;
+	new->size = i;
 	new->refs = 1;
 	*offset = *offset / SIZE_PAGE;
